+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added nullish coalescing operator (`??`)
 - Added `replace()` method
 - Added `min()` and `max()` methods
-- Added `numbers()`, `sum()`, `avg()` and `count()` methods
+- Added `numbers()`, `sum()`, `avg()`, `count()`, `percentile()` (with alias `p()`) and `median()` methods
 - Added `indexOf()` and `lastIndexOf()` methods
 - Added `toLowerCase()`, `toUpperCase()` and `trim()` methods
 - Added math methods `abs()`, `acos()`, `acosh()`, `asin()`, `asinh()`, `atan()`, `atan2()`, `atanh()`, `cbrt()`, `ceil()`, `clz32()`, `cos()`, `cosh()`, `exp()`, `expm1()`, `floor()`, `fround()`, `hypot()`, `imul()`, `log()`, `log10()`, `log1p()`, `log2()`, `pow()`, `round()`, `sign()`, `sin()`, `sinh()`, `sqrt()`, `tan()`, `tanh()` and `trunc()`. All the methods works the same as static methods of [`Math`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math) in JavaScript